[ptn] refactored code into VideoCoreMediaReceiver class
