+++ Source/VideoCoreRTC/Public/VideoCoreSignalingComponent.h
@@ -1,25 +1,24 @@
-// Fill out your copyright notice in the Description page of Project Settings.
+//
+// VideoCoreSignalingComponent.h
+//
+//  Generated on February 8 2020
+//  Copyright 2021 Regents of the University of California
+//
 
 #pragma once
 
 #include "CoreMinimal.h"
 #include "Components/ActorComponent.h"
 #include "SocketIOClientComponent.h"
-
-#pragma warning(disable:4596)
-#pragma warning(disable:4800)
-#include "mediasoupclient.hpp"
-
+#include "SIOJsonValue.h"
 #include "VideoCoreSignalingComponent.generated.h"
+
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FVideoCoreRtcSignalingConnected, USIOJsonObject*, RouterCaps);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FVideoCoreRtcSignallingDisconnected, FString, Reason);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FVideoCoreRtcSubsystemInitialized, USIOJsonObject*, MyCaps);
 
 UCLASS( ClassGroup=(Custom), meta=(BlueprintSpawnableComponent) )
 class VIDEOCORERTC_API UVideoCoreSignalingComponent : public UActorComponent
-	, public mediasoupclient::Transport::Listener
-	, public mediasoupclient::Consumer::Listener
-	, public rtc::VideoSinkInterface<webrtc::VideoFrame>
-	//, public webrtc::ObserverInterface
-	//, public webrtc::RtpReceiverObserverInterface
-	
 {
 	GENERATED_BODY()
 private:
@@ -29,6 +28,19 @@ private:
 public:	
 	// Sets default values for this component's properties
 	UVideoCoreSignalingComponent();
+
+	UFUNCTION()
+	USocketIOClientComponent* getSocketIOClientComponent() const { return sIOClientComponent; }
+
+public:
+	UPROPERTY(BlueprintAssignable)
+	FVideoCoreRtcSignalingConnected OnRtcSignalingConnected;
+
+	UPROPERTY(BlueprintAssignable)
+	FVideoCoreRtcSignallingDisconnected OnRtcSiganlingDisconnected;
+
+	UPROPERTY(BlueprintAssignable)
+	FVideoCoreRtcSubsystemInitialized OnRtcSubsystemInitialized;
 
 protected:
 	// Called when the game starts
@@ -45,28 +57,4 @@ private:
 	void onDisconnected(TEnumAsByte<ESIOConnectionCloseReason> Reason);
 
 	void setupVideoCoreServerCallbacks();
-
-	// TODO: refactor this (into a separate actor/class?)
-	void loadMediaSoupDevice(TSharedPtr<FJsonValue> rtpCapabilities);
-	void subscribe();
-	void consume(mediasoupclient::RecvTransport* t);
-
-	// mediasoupclient::Transport::Listener interface
-	std::future<void> OnConnect(
-		mediasoupclient::Transport* transport, const nlohmann::json& dtlsParameters) override;
-	void OnConnectionStateChange(
-		mediasoupclient::Transport* transport, const std::string& connectionState) override;
-
-	// mediasoupclient::Consumer::Listener interface
-	void OnTransportClose(mediasoupclient::Consumer* consumer) override;
-
-	// rtc::VideoSinkInterface<webrtc::VideoFrame> interface
-	void OnFrame(const webrtc::VideoFrame& vf);
-
-	mediasoupclient::Device device_;
-	mediasoupclient::RecvTransport* recvTransport_;
-	mediasoupclient::Consumer* consumer_;
-
-	//webrtc::MediaStreamInterface* stream_;
-	rtc::scoped_refptr<webrtc::MediaStreamInterface> stream_;
 };
